Fixed gremlin query creation error that didn't filter relationships by source node label.
