+++ tests/fixtures/minimal.yml
@@ -20,6 +20,7 @@ model:
             - name: repo
               type: String
         list: true
+        
   # Project
   - name: Project
     props:
